pulled out into searchBar component for dynamic ability
